requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.7
django-classy-tags==4.0.0
django-sekizai==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
fonttools==4.39.2
idna==3.4
importlib-resources==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
utils==1.0.1
zipp==3.15.0
